requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
catboost
PyYAML
xgboost
mlflow==2.2.2
python-box==6.0.2
ensure==1.0.2
Dill
flask
flask_cors
lime
-e .